make YAML converters and operator< of MatStencil private
