+++ src/operators/createlaplacian.cpp
@@ -19,6 +19,9 @@
 # include <petibm/type.h>
 # include <petibm/misc.h>
 
+// private functions
+# include "../private/private.h"
+
 
 namespace // anonymous namespace for internal linkage
 {
